# Optimize `CodedBufferWriter._copyInto` to memcpy